fixed mobx error; renamed a file; raised version to 1.3.0
