+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@zajno/common",
-  "version": "1.3.0-beta",
+  "version": "1.3.0",
   "description": "Zajno's re-usable utilities for JS/TS projects",
   "repository": {
     "type": "git",
