+++ src/viewModels/index.ts
@@ -7,4 +7,4 @@ export * from './wrappers';
 export * from './SelectModel';
 export * from './MultiSelectModel';
 
-export * from './PromptModalViewModel';
+export * from './PromptModal';
